backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0 